add region request scheduler
